+++ lib/src/models.dart
@@ -17,3 +17,4 @@ part 'models/song.dart';
 part 'models/video_with_comments.dart';
 part 'models/video_with_comments_list.dart';
 part 'models/video_live_info.dart';
+part 'models/serializable.dart';