+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Changed `Video` constructor parameter `duration` to optional
 - **Breaking**: All uses of `startScheduled`, `startActual`, `endActual`, and `liveViewers`
 have been replaced with `VideoLiveInfo`
+- Added `Serializable`
+- Changed `Video` to extend `Serializable`
 
 ## 0.3.1